// src/app/api/swagger/models/api-order-view-dto.ts
/* tslint:disable */
/* eslint-disable */
import { ApiCustomerContactInfoAltDto } from '../models/api-customer-contact-info';
import { ApiDealDirectionFlowEnum } from '../models/api-deal-direction-flow-enum';
import { ApiDealStateEnum } from '../models/api-deal-state-enum';
import { ApiFederalLawLimitAltEnum } from '../models/api-federal-law-limit';
import { ApiM4DDealSignRequiredEnum } from '../models/api-m-4-d-deal-sign-required-enum';
import { ApiOrderDetailDto } from '../models/api-order-detail-dto';
import { ApiOrderDocumentDto } from '../models/api-order-document-dto';
import { ApiOrderStateEnum } from '../models/api-order-state-enum';
import { ApiOrganizationCountryAltEnum } from '../models/api-organization-country';
import { ApiTradeBusinessFlowEnum } from '../models/api-trade-business-flow-enum';
import { ApiViewApplicationsForTradeDto } from '../models/api-view-applications-for-trade-dto';
import { ApiZmoFinanceSourceEnum } from '../models/api-zmo-finance-source-enum';
export interface ApiOrderViewDto {
  ApplicationId?: number;
  ApplicationSet?: ApiViewApplicationsForTradeDto | null;
  BusinessFlowType?: ApiTradeBusinessFlowEnum;
  CanSignDeal?: boolean;
  CancellationReason?: string | null;
  ContactInfo?: ApiCustomerContactInfoAltDto | null;
  CreatorUserId?: number | null;
  CurrentCustomerHasDealForThisOrder?: boolean;
  CurrentParticipantHasDealForThisOrder?: boolean;
  CurrentStageId?: number;
  CustomerFullName?: string | null;
  CustomerId?: string | null;
  CustomerInn?: string | null;
  CustomerPhysAddress?: string | null;
  DealConclusionTermInfo?: string | null;
  DealDirectionFlow?: ApiDealDirectionFlowEnum;
  DealState?: ApiDealStateEnum | null;
  DeliveryPlace?: string | null;
  DeliveryTerms?: string | null;
  Description?: string | null;
  EnableMonetizationMarket?: boolean;
  EnableMonetizationZmo?: boolean;
  ExternalContractTemplateId?: number | null;
  FederalLawLimit?: ApiFederalLawLimitAltEnum | null;
  FreeWinnerSelectionEnabled?: boolean;
  Id?: number;
  InitialPrice?: number | null;
  IsBasedOnPublicOffer?: boolean | null;
  IsContractInElectronicForm?: boolean | null;
  IsEisConcludeDealRequired?: boolean;
  IsFederalLawsEnable?: boolean;
  IsParentContractNumberNotSpecified?: boolean | null;
  IsParticipantCanCreateDeal?: boolean;
  IsSendContractAllowed?: boolean;
  LastModificationDate?: string | null;
  LotAuctionId?: number | null;
  LotId?: number;
  M4DDealSignRequired?: ApiM4DDealSignRequiredEnum;
  OrderDetails?: Array<ApiOrderDetailDto> | null;
  OrderDocuments?: Array<ApiOrderDocumentDto> | null;
  OrderName?: string | null;
  OrderState?: ApiOrderStateEnum;
  OriginTradeId?: number | null;
  ParentContractNumber?: string | null;
  ParentNotTookPlaceTradeId?: number | null;
  ParticipantComment?: string | null;
  ParticipantFullName?: string | null;
  ParticipantGuid?: string;
  ParticipantId?: string | null;
  ParticipantInn?: string | null;
  ParticipantOrganizationCountry?: ApiOrganizationCountryAltEnum;
  PlanedDealSignDate?: string | null;
  PublicationDate?: string | null;
  ReadOnly?: boolean;
  RegionCode?: string | null;
  TradeIdentificationCode?: string | null;
  ZmoFinanceSource?: ApiZmoFinanceSourceEnum | null;
  ZmoFzType?: number | null;
}